moved test from suites to tests
